wrote custom cuda kernels
wrote custom cuda kernels
performance go BRRRTTTT

(Optimized code, reduced memory usage, added safeguards against using too much mem, video quality max)
